Restructure the test and update README.md
